pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Fedor.Shurgin <shurginfedor@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^23.0.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"
requests = "^2.32.3"
lxml = "^5.3.0"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
